Mode clipboard module content to api._modules
The module only defined two commands using the api. As utils should not
depend on the api and the _modules are made for defining simple api
commands this seems like the logical approach.
